Add mock-foo.ml as example of mock module pattern
I'm still working out the pattern that we want to follow with our mock modules.
This example shows the use of polymorphic variants to tag functions, which
we'll store in a hash table. We use polymorphic variants instead of proper sum
types because we won't necessarily know the types of the functions of the
module we're mocking when we're filtering with camlp4. This problem could be
avoided by requiring a module signature in order for a module to be mocked, so
that we can iterate over the function signatures to extract types. I'm not sure
which approach is better.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>